feat: begin implementing methods for interacting with db

+++ backend/src/engine/service.rs
@@ -2,6 +2,6 @@ use crate::engine::checks::check::Check;
 
 pub struct Service {
     pub name: String,
-    pub weight: i32,
+    pub weight: i64,
     pub check: Box<dyn Check + Send + Sync>,
 }
